# api/src/main/resources/public/swagger/api.yaml
openapi: 3.0.2
info:
  title: API
  version: 1.0.0
servers:
  - url: /v1/api
security:
  - bearer-atk-token: []
...
paths:
...
  /projects:
    get:
      tags:
        - Projects
      operationId: getProjects
      description: 'Get all users projects ordered by most recent'
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDto'
    post:
      tags:
        - Projects
      operationId: addProject
      description: 'Add project by user'
      parameters:
        - in: header
          name: X-User-ID
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        description: Body of project to be created
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  type: string
                ideaImage:
                  type: string
                  format: binary
                solutionImage:
                  type: string
                  format: binary
              required:
                - data
                - ideaImage
                - solutionImage
      responses:
        '204':
          description: Returned when project creation operation is successful
...
components:
  schemas:
...
    ProjectDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        title:
          type: string
        ideaImage:
          type: string
        solutionImage:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        points:
          type: integer
        creationTimestamp:
          type: integer
          format: int64
...